changes between floor editor, now elevators/stairs are disconnected/connected to the stairwell/elevator on right click, you will get DATABASE errors if XY of elevators aren't aligned correctly
